+++ app/src/main/java/com/example/filmer/App.kt
@@ -1,11 +1,13 @@
 package com.example.filmer
 
+
 import android.app.Application
 import android.content.Context
 import com.example.filmer.di.AppComponent
 import com.example.filmer.di.DaggerAppComponent
 import com.example.filmer.di.modules.DomainModule
-import com.example.filmer.di.modules.RemoteModule
+import com.example.remote_module.DaggerRemoteComponent
+import com.example.sql_module.DaggerSqlComponent
 
 class App : Application() {
 
@@ -15,8 +17,12 @@ class App : Application() {
         super.onCreate()
         instance = this
 
+        val remoteProvider = DaggerRemoteComponent.create()
+        val sqlProvider = DaggerSqlComponent.builder().context(this).build()
+
         appComponent = DaggerAppComponent.builder()
-            .remoteModule(RemoteModule())
+            .sqlProvider(sqlProvider)
+            .remoteProvider(remoteProvider)
             .domainModule(DomainModule(this))
             .build()
     }
