+++ app/src/main/java/com/example/filmer/data/DataBase.kt
@@ -1,10 +1,8 @@
 package com.example.filmer.data
 
-import androidx.lifecycle.LiveData
 import io.reactivex.rxjava3.core.Observable
-import kotlinx.coroutines.flow.Flow
 
 interface DataBase<T> {
-    fun getFilmDB() : Observable<List<FilmData>>
+    fun getFilmDB() : Observable<List<com.example.sql_module.FilmData>>
     fun addFilms(films: List<T>)
 }
