--- app/src/main/java/com/example/filmer/data/api/FilmApi.kt
@@ -1,25 +1,0 @@
-package com.example.filmer.data.api
-
-import io.reactivex.rxjava3.core.Observable
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Path
-import retrofit2.http.Query
-
-interface FilmApi {
-    @GET("movie/{category}")
-    fun getFilms(
-        @Path("category") category: String,
-        @Query("api_key") apiKey: String,
-        @Query("language") language: String,
-        @Query("page") page: Int
-    ): Observable<ApiQResult>
-
-    @GET("search/movie")
-    fun getFilmsSearch(
-        @Query("api_key") apiKey: String,
-        @Query("language") language: String,
-        @Query("query") query: String,
-        @Query("page") page: Int
-    ): Observable<ApiQResult>
-}
