--- app/src/main/java/com/example/filmer/data/api/FilmApiConstants.kt
@@ -1,6 +1,0 @@
-package com.example.filmer.data.api
-
-object FilmApiConstants {
-    const val BASE_URL = "https://api.themoviedb.org/3/"
-    const val IMAGES_URL = "https://image.tmdb.org/t/p/"
-}
