--- app/src/main/java/com/example/filmer/data/api/FilmResult.kt
@@ -1,18 +1,0 @@
-package com.example.filmer.data.api
-
-data class FilmResult(
-    val adult: Boolean,
-    val backdrop_path: String,
-    val genre_ids: List<Int>,
-    val id: Int,
-    val original_language: String,
-    val original_title: String,
-    val overview: String,
-    val popularity: Double,
-    val poster_path: String,
-    val release_date: String,
-    val title: String,
-    val video: Boolean,
-    val vote_average: Double,
-    val vote_count: Int
-)
