+++ app/src/main/java/com/example/filmer/data/apiUtils/ResultToFilmsConverter.kt
@@ -1,0 +1,22 @@
+package com.example.filmer.data.apiUtils
+
+import com.example.sql_module.FilmData
+
+object ResultToFilmsConverter {
+    fun convertToFilmsList(list: List<com.example.remote_module.entity.FilmResult>): List<FilmData> {
+        val result = mutableListOf<FilmData>()
+
+        list.forEach {
+            result.add(
+                FilmData(
+                    poster = it.poster_path,
+                    title = it.title,
+                    description = it.overview,
+                    rating = it.vote_average,
+                    isFavorite = false,
+                )
+            )
+        }
+        return result
+    }
+}
