--- app/src/main/java/com/example/filmer/data/db/DatabaseHelper.kt
@@ -1,6 +1,0 @@
-package com.example.filmer.data.db
-
-import android.content.Context
-import android.database.sqlite.SQLiteDatabase
-import android.database.sqlite.SQLiteOpenHelper
-import javax.inject.Inject
