--- app/src/main/java/com/example/filmer/di/modules/RemoteModule.kt
@@ -1,50 +1,0 @@
-package com.example.filmer.di.modules
-
-import com.example.filmer.data.api.FilmApi
-import com.example.filmer.data.api.FilmApiConstants
-import dagger.Module
-import dagger.Provides
-import hu.akarnokd.rxjava3.retrofit.RxJava3CallAdapterFactory
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-import javax.inject.Named
-import javax.inject.Singleton
-
-@Module
-class RemoteModule {
-
-    @Singleton
-    @Provides
-    fun provideOkHttpClient(): OkHttpClient {
-        return OkHttpClient.Builder()
-            .callTimeout(8, TimeUnit.SECONDS)
-            .readTimeout(8, TimeUnit.SECONDS)
-
-            .addInterceptor(HttpLoggingInterceptor().apply {
-                level = HttpLoggingInterceptor.Level.BASIC
-            })
-
-            .build()
-    }
-
-    @Singleton
-    @Provides
-    @Named("FilmDataApi")
-    fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(FilmApiConstants.BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
-            .client(okHttpClient)
-            .build()
-    }
-
-    @Singleton
-    @Provides
-    fun provideFilmApi(@Named("FilmDataApi") retrofit: Retrofit): FilmApi =
-        retrofit.create(FilmApi::class.java)
-
-}
