+++ app/src/main/java/com/example/filmer/domain/Interact.kt
@@ -1,12 +1,11 @@
 package com.example.filmer.domain
 
-import com.example.filmer.data.FilmData
+import com.example.sql_module.FilmData
 import com.example.filmer.data.FilmDataBase
 import com.example.filmer.data.PreferenceProvider
-import com.example.filmer.data.api.FilmApi
-import com.example.filmer.data.api.FilmApiKey
-import com.example.filmer.data.api.ResultToFilmsConverter
+import com.example.filmer.data.apiUtils.ResultToFilmsConverter
 import com.example.filmer.data.db.SQLInteractor
+import com.example.remote_module.FilmApi
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.schedulers.Schedulers
 import io.reactivex.rxjava3.subjects.BehaviorSubject
@@ -23,14 +22,15 @@ class Interact @Inject constructor(
 
     var progressBarState: BehaviorSubject<Boolean> = BehaviorSubject.create()
 
-    fun getFilmDataBase(): Observable<List<FilmData>> = dbase.getFilmDB()
+    fun getFilmDataBase(): Observable<List<com.example.sql_module.FilmData>> = dbase.getFilmDB()
 
     fun searchNewFilms(query: String, onReload: Boolean = false) {
         lastApiRequest = System.currentTimeMillis()
         if (onReload) dbase.resetPage()
 
+
         val disposable = filmApi.getFilmsSearch(
-            FilmApiKey.FILM_API_KEY,
+            com.example.remote_module.entity.FilmApiKey.FILM_API_KEY,
             "ru",
             query.lowercase(),
             dbase.getLastLoadedPage("querry_$query")
@@ -39,7 +39,6 @@ class Interact @Inject constructor(
         }.subscribeOn(Schedulers.io())
             .observeOn(Schedulers.io())
             .subscribe({ data ->
-                println("search essssssssssssss")
                 if (data.isNotEmpty())
                     dbase.incrementLastLoadedPage()
 
@@ -49,7 +48,6 @@ class Interact @Inject constructor(
 
                 progressBarState.onNext(false)
             }, {
-                println("search errrrrrrrrrrrr")
                 progressBarState.onNext(false)
             })
     }
@@ -61,7 +59,7 @@ class Interact @Inject constructor(
 
         val disposable = filmApi.getFilms(
             currCategory,
-            FilmApiKey.FILM_API_KEY,
+            com.example.remote_module.entity.FilmApiKey.FILM_API_KEY,
             "ru",
             dbase.getLastLoadedPage(currCategory)
         ).map {
@@ -70,7 +68,6 @@ class Interact @Inject constructor(
             .observeOn(Schedulers.io())
             .subscribe({ data ->
 
-                println("load essssssssssssss")
                 if (onReload && data.isNotEmpty())
                     sqlInteractor.setNewFilmsListToDb(data)
                 else if (data.isNotEmpty()) {
@@ -80,37 +77,8 @@ class Interact @Inject constructor(
 
                 progressBarState.onNext(false)
             }, {
-                println("load errrr $it")
                 progressBarState.onNext(false)
             })
-
-//            .enqueue(object : Callback<ApiQResult> {
-//                override fun onResponse(call: Call<ApiQResult>, response: Response<ApiQResult>) {
-//                    try {
-//                        val data =
-//                            ResultToFilmsConverter.convertToFilmsList(response.body()?.results!!)
-//                        dbase.incrementLastLoadedPage()
-//
-//                        Completable.fromSingle<List<FilmData>> {
-//                            if (onReload && data.isNotEmpty())
-//                                sqlInteractor.setNewFilmsListToDb(data)
-//                            else if (data.isNotEmpty())
-//                                sqlInteractor.addFilmsToDb(data)
-//                            progressBarState.onNext(false)
-//                        }
-//                            .subscribeOn(Schedulers.io())
-//                            .subscribe()
-//
-//                    } catch (_: Exception) {
-//                        progressBarState.onNext(false)
-//                    }
-//                }
-//
-//                override fun onFailure(call: Call<ApiQResult>, t: Throwable) {
-//                    progressBarState.onNext(false)
-//
-//                }
-//            })
     }
 
     fun getDefCategoryFP() = preferences.getDefaultCategory()
