+++ app/src/main/java/com/example/filmer/util/FilmSearch.kt
@@ -1,14 +1,14 @@
 package com.example.filmer.util
 
 import android.util.Log
-import com.example.filmer.data.FilmData
+import com.example.sql_module.FilmData
 import com.example.filmer.views.rvadapters.RAdapter
 import javax.inject.Inject
 
 class FilmSearch @Inject constructor() {
 
     fun search(
-        filmsDataBase: List<FilmData>,
+        filmsDataBase: List<com.example.sql_module.FilmData>,
         adapter: RAdapter,
         onlyFavorites: Boolean = false,
         query: String?
@@ -16,7 +16,7 @@ class FilmSearch @Inject constructor() {
         Log.d("bebe", "newList - ${filmsDataBase.size} : adapter - ${adapter.data.size}")
 
         val list = adapter.data
-        var origin = mutableListOf<FilmData>()
+        var origin = mutableListOf<com.example.sql_module.FilmData>()
         if (onlyFavorites) {
             filmsDataBase.forEach {
                 if (it.isFavorite) origin.add(it)
