+++ app/src/main/java/com/example/filmer/views/MainActivity.kt
@@ -1,7 +1,6 @@
 package com.example.filmer.views
 
 import android.os.Bundle
-import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
@@ -9,7 +8,7 @@ import com.example.filmer.R
 import com.example.filmer.views.fragments.SelectionsFragment
 import com.example.filmer.views.fragments.TVFragment
 import com.example.filmer.views.fragments.WatchLaterFragment
-import com.example.filmer.data.FilmData
+import com.example.sql_module.FilmData
 import com.example.filmer.databinding.ActivityMainBinding
 import com.example.filmer.views.fragments.FilmDetailsFragment
 import com.example.filmer.views.fragments.SettingsFragment
@@ -122,7 +121,7 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    fun launchDetailsFragment(film: FilmData) {
+    fun launchDetailsFragment(film: com.example.sql_module.FilmData) {
         val bundle = Bundle()
         bundle.putParcelable("film", film)
         val fragment = FilmDetailsFragment()
