+++ remote_module/src/main/java/com/example/remote_module/entity/FilmApiConstants.kt
@@ -1,0 +1,6 @@
+package com.example.remote_module.entity
+
+object FilmApiConstants {
+    const val BASE_URL = "https://api.themoviedb.org/3/"
+    const val IMAGES_URL = "https://image.tmdb.org/t/p/"
+}
