+++ remote_module/src/main/java/com/example/remote_module/entity/FilmApiKey.kt
@@ -1,0 +1,5 @@
+package com.example.remote_module.entity
+
+object FilmApiKey {
+    const val FILM_API_KEY = "2a1c0c001e47c51d8578992ff132123a"
+}
