+++ sql_module/src/main/java/com/example/sql_module/SqlProvider.kt
@@ -1,0 +1,9 @@
+package com.example.sql_module
+
+import com.example.sql_module.sql.FavoritesDBDao
+import com.example.sql_module.sql.FilmDBDao
+
+interface SqlProvider {
+    fun provideFilmDBDao(): FilmDBDao
+    fun provideFavoritesDBDao(): FavoritesDBDao
+}
